scripts/requirements.txt
playwright>=1.57.0
pillow>=12.0.0

